+++ codex_of_the_damned/static/js/statistics.js
@@ -1,5 +1,4 @@
 function binom(n, k) {
-    console.log(`binom ${n}, ${k}`)
     let res = 1
     for (const i of Array(Math.min(k, n - k)).keys()) {
         res *= (n - i) / (i + 1)
@@ -8,7 +7,6 @@ function binom(n, k) {
 }
 
 function atLeastOne(stack, copies, draws) {
-    console.log(`aLO ${stack}, ${copies}, ${draws}`)
     let res = 0
     const denominator = binom(stack, draws)
     for (const i of Array(Math.min(copies, draws)).keys()) {
@@ -20,7 +18,6 @@ function atLeastOne(stack, copies, draws) {
 }
 
 function howManyNeeded(stack, draws, expectation) {
-    console.log(`hMN ${stack}, ${draws}, ${expectation}`)
     for (const copies of Array(stack).keys()) {
         const probability = atLeastOne(stack, copies + 1, draws)
         if (probability >= expectation) {